backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.13",
    "numpy>=2.3.0",
    "openai>=1.88.0",
    "pydantic-settings>=2.9.1",
    "pypdf>=5.1.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
